+++ patches/src/main/kotlin/app/morphe/patches/youtube/interaction/seekbar/SeekbarThumbnailPreviewPatch.kt
@@ -17,9 +17,6 @@ import app.morphe.patches.youtube.misc.extension.sharedExtensionPatch
 import app.morphe.patches.youtube.misc.settings.PreferenceScreen
 import app.morphe.patches.youtube.misc.settings.settingsPatch
 import app.morphe.patches.youtube.shared.Constants.COMPATIBILITY_YOUTUBE
-import app.morphe.patches.youtube.shared.SeekbarBigBoardsUpdateFingerprint
-import app.morphe.patches.youtube.shared.SeekbarFineScrubbingBitmapFingerprint
-import app.morphe.patches.youtube.shared.SeekbarTrackballPosXAndTimeMillisFingerprint
 import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
 import com.android.tools.smali.dexlib2.iface.instruction.TwoRegisterInstruction
